[DEV] Made sure the name generator reads learnset from a file

+++ include/qg_random.hpp
@@ -5,4 +5,4 @@
 void rand_seed(int64_t seed);
 float rand_float01();
 
-int rand_weighted_index(int8_t *weights, int num_items);
+int rand_weighted_index(int32_t *weights, int num_items);
